+++ screens/HomeScreen/index.tsx
@@ -1,0 +1,26 @@
+import { StyleSheet, FlatList } from "react-native";
+
+import { View } from "@/components/Themed";
+import categories from "@/assets/data/categories";
+import HomeCategory from "@/components/HomeCategory";
+
+export default function HomeScreen() {
+  return (
+    <View style={styles.container}>
+      <FlatList
+        data={categories.items}
+        renderItem={({ item }) => <HomeCategory category={item} />}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: "bold",
+  },
+});
