+++ screens/MovieDetailsScreen/index.tsx
@@ -1,0 +1,15 @@
+import { Text, StyleSheet, View } from "react-native";
+
+export default function MovieDetailsScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>dehfkejf</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
